Handle multiple current sources
It seems that some packages are current in both testing and unstable,
which means if one of those is the version I'm targeting, both will be
reported as valid sources for a given package version, and I can't
assume I have something from an unexpected source just because I don't
expect _all_ the possible sources.

Rewrite to handle the possibility of multiple sources for a single
current package version, and only report packages where none of the
current sources look reasonable.